# Improve overall performance of request handling