--- a/src/Thinktecture.Relay.Connector/Targets/ClientRequestHandler.cs
+++ b/src/Thinktecture.Relay.Connector/Targets/ClientRequestHandler.cs
@@ -14,11 +14,9 @@ namespace Thinktecture.Relay.Connector.Targets;
 
 /// <inheritdoc />
 public partial class ClientRequestHandler<TRequest, TResponse, TAcknowledge> : IClientRequestHandler<TRequest>
-// ReSharper disable RedundantNameQualifier; (this is needed in 6.0, see https://github.com/dotnet/runtime/issues/58550)
-	where TRequest : Thinktecture.Relay.Transport.IClientRequest
-	where TResponse : Thinktecture.Relay.Transport.ITargetResponse, new()
-	where TAcknowledge : Thinktecture.Relay.Acknowledgement.IAcknowledgeRequest, new()
-// ReSharper restore RedundantNameQualifier
+	where TRequest : IClientRequest
+	where TResponse : ITargetResponse, new()
+	where TAcknowledge : IAcknowledgeRequest, new()
 {
 	private readonly Uri _acknowledgeEndpoint;
 	private readonly IAcknowledgeTransport<TAcknowledge> _acknowledgeTransport;