--- a/src/Thinktecture.Relay.Server.Protocols.RabbitMq/TenantHandler.cs
+++ b/src/Thinktecture.Relay.Server.Protocols.RabbitMq/TenantHandler.cs
@@ -84,6 +84,7 @@ public partial class TenantHandler<TRequest, TAcknowledge> : ITenantHandler, IDi
 			Log.CouldNotParseAcknowledge(_logger, acknowledgeId);
 		}
 	}
+
 	private async Task ConsumerReceivedAsync(BasicDeliverEventArgs @event)
 	{
 		var request = JsonSerializer.Deserialize<TRequest>(@event.Body.Span) ??