--- a/src/Thinktecture.Relay.Server.Protocols.RabbitMq/TenantTransport.Log.cs
+++ b/src/Thinktecture.Relay.Server.Protocols.RabbitMq/TenantTransport.Log.cs
@@ -15,5 +15,13 @@ public partial class TenantTransport<TRequest, TAcknowledge>
 			"An error occured while dispatching request {RelayRequestId} to tenant {TenantName} queue")]
 		public static partial void ErrorDispatchingRequest(ILogger logger, Exception ex, Guid relayRequestId,
 			string tenantName);
+
+		[LoggerMessage(LoggingEventIds.TenantTransportLoadedTenantNames, LogLevel.Information,
+			"Loaded {TenantCount} tenant names")]
+		public static partial void LoadedTenantNames(ILogger logger, int tenantCount);
+
+		[LoggerMessage(LoggingEventIds.TenantTransportEnsuredTenantQueue, LogLevel.Trace,
+			"Ensured queue {QueueName} for tenant {TenantName}")]
+		public static partial void EnsuredTenantQueue(ILogger logger, string queueName, string tenantName);
 	}
 }