… src/docker/Thinktecture.Relay.Connector.Docker/appsettings.Development.json
… src/docker/Thinktecture.Relay.Connector.Docker/appsettings.json