This PR improves the overall performance. There were two major performance flaws which will be addressed:

1. Using `ThreadPool.QueueUserWorkItem` harms the possibility of concurrent requests. The way to go is to execute the asynchronous method without an `await`.
2. Ensure the RabbitMQ queues kill the model on every publish in a pressure scenario. This must be done in front (e.g., at the start of the process or the first connection of a new tenant).

A lesser but equally important fact was that we updated the tenant info cache not only once but multiple times in case of concurrent requests for one tenant. This resulted in multiple database requests and exhausted the connection pool under load scenarios.

From initially approx. 20 requests/second, the load testing results now into the following:

```

          /\      |‾‾| /‾‾/   /‾‾/
     /\  /  \     |  |/  /   /  /
    /  \/    \    |     (   /   ‾‾\
   /          \   |  |\  \ |  (‾)  |
  / __________ \  |__| \__\ \_____/ .io

     execution: local
        script: ./test.js
        output: -

     scenarios: (100.00%) 1 scenario, 100 max VUs, 1m0s max duration (incl. graceful stop):
              * relay: 100 looping VUs for 30s (gracefulStop: 30s)


     data_received..................: 49 MB  1.6 MB/s
     data_sent......................: 5.5 MB 183 kB/s
     http_req_blocked...............: avg=7.13µs  min=0s     med=1µs     max=4.75ms   p(90)=2µs     p(95)=3µs
     http_req_connecting............: avg=5.17µs  min=0s     med=0s      max=3.23ms   p(90)=0s      p(95)=0s
     http_req_duration..............: avg=59.43ms min=1.58ms med=57.63ms max=157.68ms p(90)=68.81ms p(95)=73.15ms
       { expected_response:true }...: avg=59.43ms min=1.58ms med=57.63ms max=157.68ms p(90)=68.81ms p(95)=73.15ms
     http_req_failed................: 0.00%  ✓ 0           ✗ 50469
     http_req_receiving.............: avg=35.28µs min=7µs    med=28µs    max=13.3ms   p(90)=42µs    p(95)=52µs
     http_req_sending...............: avg=11.17µs min=1µs    med=6µs     max=15.65ms  p(90)=11µs    p(95)=13µs
     http_req_tls_handshaking.......: avg=0s      min=0s     med=0s      max=0s       p(90)=0s      p(95)=0s
     http_req_waiting...............: avg=59.38ms min=1.54ms med=57.59ms max=157.65ms p(90)=68.75ms p(95)=73.09ms
     http_reqs......................: 50469  1679.395374/s
     iteration_duration.............: avg=59.48ms min=1.61ms med=57.67ms max=157.7ms  p(90)=68.85ms p(95)=73.21ms
     iterations.....................: 50469  1679.395374/s
     vus............................: 100    min=100       max=100
     vus_max........................: 100    min=100       max=100


running (0m30.1s), 000/100 VUs, 50469 complete and 0 interrupted iterations
relay ✓ [======================================] 100 VUs  30s
```

```
This is ApacheBench, Version 2.3 <$Revision: 1903618 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking localhost (be patient).....done


Server Software:        gunicorn/19.9.0
Server Hostname:        localhost
Server Port:            5010

Document Path:          /trace/TestTenant1/httpbin/get
Document Length:        337 bytes

Concurrency Level:      100
Time taken for tests:   1.928 seconds
Complete requests:      3000
Failed requests:        0
Total transferred:      2921608 bytes
HTML transferred:       1011000 bytes
Requests per second:    1555.99 [#/sec] (mean)
Time per request:       64.268 [ms] (mean)
Time per request:       0.643 [ms] (mean, across all concurrent requests)
Transfer rate:          1479.81 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.6      0       4
Processing:     2   62  12.7     61     117
Waiting:        2   62  12.7     61     117
Total:          2   63  12.6     61     121

Percentage of the requests served within a certain time (ms)
  50%     61
  66%     66
  75%     70
  80%     72
  90%     78
  95%     82
  98%     85
  99%     91
 100%    121 (longest request)
```

This load test was set up in the Docker environment without any other changes, with only one server and connector.